File restructure, added subfolders to frontend src.
