+++ backend/src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.data.mongodb.uri=${MONGODB_URI}
+spring.data.mongodb.uri=mongodb+srv://speakfluid:speak@cluster0.c3onz4y.mongodb.net/test
 spring.data.mongodb.database=mytranscripts
