+++ frontend/src/App.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import logo from './logo.svg';
-import './App.css';
+import './styles/App.css';
 
 type Transcript = {
   turnID: string,
@@ -41,7 +40,6 @@ function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
         <input 
           onChange={handleChange}
         ></input>
